moved files to `lib` folder
