+++ src/Lib/ToolResolverInterface.php
@@ -14,14 +14,9 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StaticResolverBundle\Tool;
+namespace Pimcore\Bundle\StaticResolverBundle\Lib;
 
-use Pimcore\Tool;
-
-class ToolResolver implements ToolResolverInterface
+interface ToolResolverInterface
 {
-    public function getValidLanguages(): array
-    {
-        return Tool::getValidLanguages();
-    }
+    public function getValidLanguages(): array;
 }
